collector/requirements.txt
httpx
pika
python-dotenv